# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hackathon2026 Backend API
  version: 1.3.0
  description: |
    OpenAPI contract for the Hackathon2026 backend.

    Authentication model:
    - `POST /auth/psychologists/login` sets `session_id` and `csrf_token` cookies.
    - `POST /auth/admin/login` sets `admin_session_id` and `csrf_token` cookies.
    - Protected `GET` endpoints require the corresponding session cookie.
    - Protected non-`GET` endpoints also require header `X-CSRF-Token` with the same value as the `csrf_token` cookie.

    Practical notes for frontend integration:
    - Base URL in local development is `http://localhost:8080`.
    - Error responses use the unified shape `{ "message": "...", "field_errors": { ... } }`.
    - Browser CORS preflight is supported for configured `ALLOWED_ORIGINS`, including credentials and the `X-CSRF-Token` header.
    - Psychologist user payloads return `portal_access_until` and `blocked_until` as `null` when they are not set.
    - Psychologist user payloads may include computed `account_status` and `subscription_status`.
    - `GET /psychologists/tests` returns activity metrics such as started/completed counts and last activity timestamps for the “Мои опросники” dashboard.
    - Test creation defaults `status` to `published` when omitted.
    - Test update defaults `status` to `draft` when omitted.
    - Public test start requires `respondent_name` and `respondent_phone`.
...
    - In-progress public sessions expire after 1 hour of inactivity.
...
servers:
...
paths:
...
  /psychologists/tests/{id}/questions:
    parameters:
      - $ref: '#/components/parameters/TestIdParam'
    post:
      tags:
        - Questions
      summary: Create question
      operationId: createPsychologistQuestion
      security:
        - psychologistSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestionInput'
      responses:
        '201':
          description: Question created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Questions
      summary: List questions
      operationId: listPsychologistQuestions
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Questions for the selected test
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/tests/{id}/questions/{questionId}:
    parameters:
      - $ref: '#/components/parameters/TestIdParam'
      - $ref: '#/components/parameters/QuestionIdParam'
    get:
      tags:
        - Questions
      summary: Get question
      operationId: getPsychologistQuestion
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Question details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Questions
      summary: Update question
      operationId: updatePsychologistQuestion
      security:
        - psychologistSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuestionInput'
      responses:
        '200':
          description: Updated question
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Questions
      summary: Delete question
      operationId: deletePsychologistQuestion
      security:
        - psychologistSession: []
          csrfToken: []
      responses:
        '200':
          description: Question deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/tests/{id}/publish:
    parameters:
      - $ref: '#/components/parameters/TestIdParam'
    post:
      tags:
        - Tests
      summary: Publish test
      operationId: publishPsychologistTest
      security:
        - psychologistSession: []
          csrfToken: []
      responses:
        '200':
          description: Test published and public URL returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishTestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/tests/{id}/results:
    parameters:
      - $ref: '#/components/parameters/TestIdParam'
    get:
      tags:
        - Tests
      summary: List test submissions
      operationId: listPsychologistTestSubmissions
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Public test submissions for the selected test
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PsychologistTestSubmission'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/tests/{id}/results/{sessionId}:
    parameters:
      - $ref: '#/components/parameters/TestIdParam'
      - $ref: '#/components/parameters/SubmissionSessionIdParam'
    get:
      tags:
        - Tests
      summary: Get test submission
      operationId: getPsychologistTestSubmission
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Test submission with answers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PsychologistTestSubmission'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /public/tests/{slug}:
    parameters:
      - $ref: '#/components/parameters/SlugParam'
    get:
      tags:
        - Public Tests
      summary: Get public test
      operationId: getPublicTest
      responses:
        '200':
          description: Public test with questions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicTest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /public/tests/{slug}/start:
    parameters:
      - $ref: '#/components/parameters/SlugParam'
    post:
      tags:
        - Public Tests
      summary: Start public test
      description: |
        Creates a public session after collecting respondent data.
        `respondent_name` and `respondent_phone` are always required.
        Additional personal fields become required when the corresponding `collect_respondent_*` flag is enabled on the test.
      operationId: startPublicTest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartPublicTestInput'
      responses:
        '200':
          description: Existing in-progress session resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartPublicTestResponse'
        '201':
          description: New public session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartPublicTestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /public/tests/{slug}/progress:
    parameters:
      - $ref: '#/components/parameters/SlugParam'
    post:
      tags:
        - Public Tests
      summary: Save public test progress
      description: |
        Saves partial answers for an in-progress public test session and extends its lifetime for another hour.
        This endpoint does not require all required questions to be answered.
      operationId: savePublicTestProgress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitPublicTestInput'
      responses:
        '200':
          description: Progress saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitPublicTestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /public/tests/{slug}/submit:
    parameters:
      - $ref: '#/components/parameters/SlugParam'
    post:
      tags:
        - Public Tests
      summary: Submit public test
      description: |
        Submits or updates answers for a public test session.
        Answer payload depends on `question_type`:
        - `text`: use `answer_text`
        - `number`, `single_choice`, `scale`: use `answer_value`
        - `multiple_choice`: use `answer_values`
      operationId: submitPublicTest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitPublicTestInput'
      responses:
        '200':
          description: Test answers accepted and session completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitPublicTestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    psychologistSession:
      type: apiKey
      in: cookie
      name: session_id
      description: Psychologist session cookie returned by the psychologist login endpoint.
    adminSession:
      type: apiKey
      in: cookie
      name: admin_session_id
      description: Admin session cookie returned by the admin login endpoint.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: "Must match the `csrf_token` cookie value on protected non-GET requests."
  parameters:
    TestIdParam:
      name: id
      in: path
      required: true
      description: Test identifier
      schema:
        type: integer
        format: int64
        minimum: 1
    QuestionIdParam:
      name: questionId
      in: path
      required: true
      description: Question identifier
      schema:
        type: integer
        format: int64
        minimum: 1
...
    SubmissionSessionIdParam:
      name: sessionId
      in: path
      required: true
      description: Public test session identifier
      schema:
        type: integer
        format: int64
        minimum: 1
...
    SlugParam:
      name: slug
      in: path
      required: true
      description: Public test slug
      schema:
        type: string
        minLength: 1
...
  responses:
    BadRequest:
      description: Invalid request payload or path parameter.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, or expired session credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated request is not allowed or fails CSRF/origin checks.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - message
        - field_errors
      properties:
        message:
          type: string
          example: Validation failed
        field_errors:
          type: object
          nullable: true
          additionalProperties:
            type: string
          example:
            email: Email already exists
    StatusResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok
...
    QuestionOption:
      type: object
      required:
        - id
        - question_id
        - label
        - value
        - order_number
        - score
      properties:
        id:
          type: integer
          format: int64
        question_id:
          type: integer
          format: int64
        label:
          type: string
        value:
          type: string
        order_number:
          type: integer
          minimum: 0
        score:
          type: number
          format: double
    QuestionOptionInput:
      type: object
      additionalProperties: false
      required:
        - label
        - value
      properties:
        label:
          type: string
          minLength: 1
        value:
          type: string
          minLength: 1
        order_number:
          type: integer
          minimum: 0
        score:
          type: number
          format: double
    Question:
      type: object
      required:
        - id
        - test_id
        - text
        - question_type
        - order_number
        - is_required
      properties:
        id:
          type: integer
          format: int64
        test_id:
          type: integer
          format: int64
        text:
          type: string
        question_type:
          type: string
          enum:
            - single_choice
            - multiple_choice
            - scale
            - text
            - number
        order_number:
          type: integer
          minimum: 0
        is_required:
          type: boolean
        scale_weights:
          type: object
          additionalProperties:
            type: number
            format: double
          description: "Per-scale coefficients for vocational guidance scoring. Allowed keys: analytic, creative, social, organizer, practical."
        options:
          type: array
          items:
            $ref: '#/components/schemas/QuestionOption'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateQuestionInput:
      type: object
      additionalProperties: false
      required:
        - text
        - question_type
      properties:
        text:
          type: string
          minLength: 1
        question_type:
          type: string
          enum:
            - single_choice
            - multiple_choice
            - scale
            - text
            - number
        order_number:
          type: integer
          description: "When omitted or `0`, the backend appends the question to the end."
          minimum: 0
        is_required:
          type: boolean
        scale_weights:
          type: object
          additionalProperties:
            type: number
            format: double
          description: "Per-scale coefficients for vocational guidance scoring. Allowed keys: analytic, creative, social, organizer, practical."
        options:
          type: array
          items:
            $ref: '#/components/schemas/QuestionOptionInput'
    UpdateQuestionInput:
      type: object
      additionalProperties: false
      required:
        - text
        - question_type
      properties:
        text:
          type: string
          minLength: 1
        question_type:
          type: string
          enum:
            - single_choice
            - multiple_choice
            - scale
            - text
            - number
        order_number:
          type: integer
          description: "When omitted or `0`, the backend keeps the current order."
          minimum: 0
        is_required:
          type: boolean
        scale_weights:
          type: object
          additionalProperties:
            type: number
            format: double
          description: "Per-scale coefficients for vocational guidance scoring. Allowed keys: analytic, creative, social, organizer, practical."
        options:
          type: array
          items:
            $ref: '#/components/schemas/QuestionOptionInput'
    PublishTestResponse:
      type: object
      required:
        - test_id
        - public_slug
        - public_url
        - is_public
        - status
      properties:
        test_id:
          type: integer
          format: int64
        public_slug:
          type: string
        public_url:
          type: string
        is_public:
          type: boolean
        status:
          type: string
          enum:
            - draft
            - published
...
    PublicTest:
      type: object
      required:
        - id
        - slug
        - title
        - description
        - has_participant_limit
        - collect_respondent_age
        - collect_respondent_gender
        - collect_respondent_education
        - questions
      properties:
        id:
          type: integer
          format: int64
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        recommended_duration:
          type: integer
          minimum: 0
        max_participants:
          type: integer
          description: Maximum number of respondents allowed via the public link when `has_participant_limit` is `true`.
          minimum: 0
        has_participant_limit:
          type: boolean
          description: Whether the public link is restricted to a fixed number of respondents.
        collect_respondent_age:
          type: boolean
        collect_respondent_gender:
          type: boolean
        collect_respondent_education:
          type: boolean
        questions:
          type: array
          items:
            $ref: '#/components/schemas/PublicQuestion'
...
    StartPublicTestInput:
      type: object
      additionalProperties: false
      required:
        - respondent_name
        - respondent_phone
      properties:
        respondent_name:
          type: string
          description: Full name. Must contain at least first name and last name.
        respondent_phone:
          type: string
          description: Phone number. Stored in normalized format and must be unique per test.
        respondent_email:
          type: string
          format: email
        respondent_age:
          type: integer
          minimum: 1
          maximum: 120
        respondent_gender:
          type: string
        respondent_education:
          type: string
    StartPublicTestResponse:
      type: object
      required:
        - session
        - test
      properties:
        session:
          $ref: '#/components/schemas/PublicTestSession'
        test:
          $ref: '#/components/schemas/PublicTest'
        answers:
          type: array
          items:
            $ref: '#/components/schemas/PublicTestAnswer'
        resumed:
          type: boolean
...
    SubmitPublicTestInput:
      type: object
      additionalProperties: false
      required:
        - access_token
        - answers
      properties:
        access_token:
          type: string
          minLength: 1
        answers:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PublicAnswerInput'
    PublicTestAnswer:
      type: object
      required:
        - id
        - session_id
        - question_id
      properties:
        id:
          type: integer
          format: int64
        session_id:
          type: integer
          format: int64
        question_id:
          type: integer
          format: int64
        answer_text:
          type: string
        answer_value:
          type: string
        answer_values:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SubmitPublicTestResponse:
      type: object
      required:
        - session_id
        - status
        - answers
      properties:
        session_id:
          type: integer
          format: int64
        status:
          type: string
          enum:
            - in_progress
            - completed
        answers:
          type: array
          items:
            $ref: '#/components/schemas/PublicTestAnswer'
    PsychologistTestSubmission:
      type: object
      required:
        - session_id
        - test_id
        - status
        - started_at
        - answers_count
      properties:
        session_id:
          type: integer
          format: int64
        test_id:
          type: integer
          format: int64
        respondent_name:
          type: string
        respondent_phone:
          type: string
        respondent_email:
          type: string
          format: email
        respondent_age:
          type: integer
          minimum: 1
          maximum: 120
        respondent_gender:
          type: string
        respondent_education:
          type: string
        status:
          type: string
          enum:
            - in_progress
            - completed
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        answers_count:
          type: integer
          minimum: 0
        answers:
          type: array
          items:
            $ref: '#/components/schemas/PublicTestAnswer'
        career_result:
          $ref: '#/components/schemas/CareerResult'
...
    CareerProfessionResult:
      type: object
      required:
        - profession
        - score
      properties:
        profession:
          type: string
        score:
          type: number
          format: double
    CareerResult:
      type: object
      required:
        - scales
        - top_scales
        - top_professions
      properties:
        scales:
          type: array
          items:
            $ref: '#/components/schemas/CareerScaleResult'
        top_scales:
          type: array
          items:
            $ref: '#/components/schemas/CareerScaleResult'
        top_professions:
          type: array
          items:
            $ref: '#/components/schemas/CareerProfessionResult'
...